Refactor codebase (#27)
* Update composer.json with LaravelWorkOS package namespace and keywords

* Add new exceptions

* feat: Add DateParser class for parsing timestamps

The commit adds a new `DateParser` class in the `NjoguAmos\LaravelWorkOS\Services` namespace. This class is responsible for parsing timestamps and converting them to the desired timezone. It includes a `parse` method that takes a timestamp as input and returns a `CarbonImmutable` object.

The `DateParser` class has a constructor that accepts a boolean parameter `$convertToAppTimezone`. If this parameter is set to `true`, the parsed timestamp will be converted to the application timezone defined in the `app.timezone` configuration. Otherwise, it will be converted to the default WorkOS timezone, which is UTC.

The commit also includes a unit test file `DateParserTest.php` that tests the functionality of the `DateParser` class. It includes two test cases: one for parsing timestamps to the UTC timezone and another for parsing timestamps to the application timezone.

This commit addresses the need for a centralized date parsing mechanism in the Laravel WorkOS package and improves the flexibility of handling timestamps in different timezones.

* Add DateParser class for parsing timestamps

* Update application namespace

* Update namespace

* Refactor UserDTO to UserData

* Refactor namespace and file names for consistency

* Refactor namespace and file names for consistency

* Update UserFactory.php

* Update DateParser.php

* Update DateParserTest.php

* Update GetUserRequestTest.php